Setup tests provided by Alchemy core
The tests are run against the current alchemy master from github.

The tests are failing for alchemy versions before this PR was merged:
https://github.com/AlchemyCMS/alchemy_cms/pull/1269/commits/b831b81dbceb4ef24c3f0104aada6deacbeed731

It doesn't mean this gem is not compatible with older versions, but the shared essence tests in Alchemy were not perfectly compatible with external essences at the time.

- [x] gem structure
- [x] introduce EssenceColor to Alchemy
- [x] added basic tests for essence
- [ ] basic implementation of EssenceColor
- [ ] add tests for the behavior of EssenceColor
- [ ] add jQuery color picker plugin
- [ ] add settings for EssenceColor / picker